requirements/base.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic
pydantic>=2.8.0
pydantic-settings>=2.1.0

# Extraction PDF
pdfplumber==0.10.3
PyMuPDF==1.23.8

# Extraction Office
openpyxl==3.1.2
python-docx==1.1.0

# OCR et images
pytesseract==0.3.10
Pillow==10.1.0

# NLP
spacy>=3.8.0

# Base de données
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
asyncpg==0.29.0
aiosqlite==0.19.0

# Utilitaires
aiofiles==23.2.1
python-dotenv==1.0.0
python-json-logger==2.0.7

